+++ src/ui/processTree/windowsProcessTree.ts
@@ -2,11 +2,10 @@
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
 
-import { join } from 'path';
-import { BaseProcessTree } from './baseProcessTree';
-import { IProcess } from './processTree';
+import { getWinUtils } from '../../common/win32Utils';
+import { IProcess, IProcessTree } from './processTree';
 
-export class WindowsProcessTree extends BaseProcessTree {
+export class WindowsProcessTree implements IProcessTree {
   /**
    * @inheritdoc
    */
@@ -17,52 +16,19 @@ export class WindowsProcessTree extends BaseProcessTree {
   /**
    * @inheritdoc
    */
-  protected createProcess() {
-    const wmic = join(process.env['WINDIR'] || 'C:\\Windows', 'System32', 'wbem', 'WMIC.exe');
-    return this.spawn(wmic, [
-      'process',
-      'get',
-      'CommandLine,CreationDate,ParentProcessId,ProcessId',
-    ]);
-  }
+  async lookup<T>(onEntry: (process: IProcess, accumulator: T) => T, initial: T): Promise<T> {
+    const win = await getWinUtils();
+    for (const proc of win.getProcessInfo()) {
+      const argsStart = proc.commandLine.indexOf('" ');
+      initial = onEntry({
+        args: argsStart !== -1 ? proc.commandLine.slice(argsStart + 2) : '',
+        command: argsStart !== -1 ? proc.commandLine.slice(1, argsStart + 1) : proc.commandLine,
+        date: proc.creationDate * 1000,
+        pid: proc.processId,
+        ppid: proc.parentProcessId,
+      }, initial);
+    }
 
-  /**
-   * @inheritdoc
-   */
-  protected createParser(): (line: string) => IProcess | void {
-    // attributes columns are in alphabetic order!
-    const CMD_PAT = /^(.*)\s+([0-9]+)\.[0-9]+[+-][0-9]+\s+([0-9]+)\s+([0-9]+)$/;
-
-    return line => {
-      const matches = CMD_PAT.exec(line.trim());
-      if (!matches || matches.length !== 5) {
-        return;
-      }
-
-      const pid = Number(matches[4]);
-      const ppid = Number(matches[3]);
-      const date = Number(matches[2]);
-      let args = matches[1].trim();
-      if (!isNaN(pid) && !isNaN(ppid) && args) {
-        let command = args;
-        if (args[0] === '"') {
-          const end = args.indexOf('"', 1);
-          if (end > 0) {
-            command = args.substr(1, end - 1);
-            args = args.substr(end + 2);
-          }
-        } else {
-          const end = args.indexOf(' ');
-          if (end > 0) {
-            command = args.substr(0, end);
-            args = args.substr(end + 1);
-          } else {
-            args = '';
-          }
-        }
-
-        return { pid, ppid, command, args, date };
-      }
-    };
+    return initial;
   }
 }